chore: rename notification/notification.ts to index.ts
